adds cf-expandables
cf-expandables is included in EditsDetail.jsx and initialized, per id
the id is created in EditsGrouped.jsx and passed as a prop to EditsDetail.jsx
we can't rely on the cf-expandables initialization because the components are rendered dynamically
we rely on the id (instead of the .expandables class) because the initialization would happen mulitple times if we did. if we relied on the class the toggle function would happen multiple times per click.