Checkbox for IRS
Adds in the more detailed description of the IRS report.
Adds in the checkbox for verifying (or un-verifying) the IRS report. When this happens a parent component function is called in order to update the state, specifically the status code. This allows us to show/hide the signature component. The signature component should NOT render if the IRS has not been verified.